Added a condition to check whether the site is run in production mode.
If the site is run in production mode, the Auth0 and backend variable values are taken from environment variables (such as those defined on Heroku). If the site is run in development mode, these values are taken from environment.ts. The mode is changed by Gulp upon calling scriptsDist (vs the regular scripts task for development).