Changed the way fetching from IDB changes the page.
Before, upon fetching from IDB the isIDBResolved variable would turn to true and thus allow the relevant component to appear. This introduced an issue when IDB did not have the requested data. Now, isIDBResolved turns true only if the data exists, and if it does not, once the data is fetched from the server.