Added an indicator as to whether notifications are refreshed silently.
This is done to prevent notifications from disappearing before users can
see them (upon an auto-refresh of the notifications list, rather than when
the user opens the notification tab). This way, when the user opens the
notifications tab, it's referred to as 'non-silent refresh', which will trigger
changing the user's 'last viewed date' in the backend. When the app auto-refreshes
notifications list, it's done silently, and so it won't change the user's 'last viewed date'.