Changed admin dashboard reports/blocks pages block checks.
Since there are tests specifically for checking the blocks are 
calculated correctly, it's not necessary to run the time checks in the 
funcations checking block() and blockUser(). The important thing for 
these two functions is that they're passing on a date. The tests were 
adjusted to take it into account.