Changed the precision in AdminDashboard tests
Jasmine's default precision is 2 decimal places, which could cause tests to break due to the time passing between the test's start and the call to the function being tested. In some cases, that time was enough to change the value of the second decimal place, which then caused the test to fail. This should fix that issue.