Simplified the way the navigation menu's class is set.
Previously, the classes of the navigation menu and the button that opens
it were set manually. Now, the classes are set dynamtically based on
whether the menu should be shown and whether the button should be
shown.
This change also removes the unnecessary ngAfterViewChecked
update, which caused the NG0100 (expression has changed
after it was checked error).