Moved the posts' class-setting code to Post component.
Previously, in order to ensure only one menu is open at once (when
there isn't enough space to show the full menu), the parent components
dealt with keeping track of the open menu and closing the menus that
shouldn't be open. This caused unnecessary duplication as the
full list, main page and search results all had to set and unset the
classes for the menus of all posts currently present in the DOM.
Now, the post component deals with setting its own menus' classes,
and the currently open menu is kept in the posts service.