Added a refactored queryPosts method to SWManager.
The original queryPosts method required specifying a target and harcoding
the index, count and filters in the method. This is unscalable and
resulted in repeated code. Instead, the responsibility for setting these
parameters (e.g., index, count) belongs to the calling service, while
the ServiceWorker manager simply fetches the data (which should be
its only responsibility).