Replaced postsService variables in mainPage with internal variables.
Since we're not using the posts we keep globally in the service, there's
no need to keep them there. The previous structure added
complication as it dealt with both the main page and the full list,
which have different variables and data. This way, the main page
holds its own data, and handles the complication of waiting for two
IDB promises to resolve before running the network-based fetch.