Changed the way the menu size is checked.
Previously, we relied on element sizes, but since it took time to update when the menu was hidden, this resulted in the menu being shown if the size was checked while the menu was hidden. Now, we use the same method of calculation we use in posts, which relies on the styling and the number of buttons shown in the menu.