chore: Refactored the ItemDeleteForm (#2062)
* Added new iputs to the ItemDeleteForm.

* Refactored the deleteSingleItem method to use the new inputs.

This places the responsibility of setting the delete URL and IDB store to delete from on the parent component, as it should've been in the first place.

* Refactored the deleteMultipleItems method to use the new inputs.

This places the responsibility of setting the delete URL and IDB store to delete from on the parent component, as it should've been in the first place.

* Deleted the old inputs to the item delete form.

* Moved the thread messages delete to the Thread component.

* Unified the two (non-admin) delete methods.

* Updated the tests of the components that use the item delete form.

* Updated the ItemDeleteForm's tests.

* Moved the report closing from the ItemDeleteForm to the AdminReports component.

It should be the responsibility of the component dealing with reports - AdminReports - rather than the generic component dealing with deleting items.

* Added a missing test ot the messages component.

* Removed the 'delete all' button when looking at a thread.

If a user wants to delete the thread, they can do it from the threads view.

* Added changelog.